IDVA6-2271 Unit tests for services refactored
