+++ package.json
@@ -12,8 +12,8 @@
         "build:assets": "gulp build",
         "build:server": "tsc && cp -r src/views dist/ && cp -r locales dist/",
         "build": "npm run build:assets && npm run build:server",
-        "lint": "eslint 'src/*.ts' 'src/**/*.ts' 'test/src/**/*.ts'",
-        "lint:fix": "eslint 'src/*.ts' 'src/**/*.ts' 'test/src/**/*.ts' --fix",
+        "lint": "eslint 'src/*.ts' 'src/**/*.ts' 'test/**/*.ts'",
+        "lint:fix": "eslint 'src/*.ts' 'src/**/*.ts' 'test/**/*.ts' --fix",
         "test:coverage": "npm run test -- --coverage",
         "test": "jest",
         "sonarqube-base-branch": "sonar-scanner",
