+++ test/unit/src/services/apiClientService.test.ts
@@ -1,10 +1,11 @@
 import { Request } from "express";
 import { createPrivateApiClient } from "private-api-sdk-node";
 import PrivateApiClient from "private-api-sdk-node/dist/client";
-import { createOauthPrivateApiClient } from "../../../src/services/apiClientService";
+import { createOauthPrivateApiClient } from "../../../../src/services/apiClientService";
 import { IHttpClient } from "@companieshouse/api-sdk-node";
 import { Session } from "@companieshouse/node-session-handler";
 jest.mock("private-api-sdk-node");
+jest.mock("../../../../src/lib/Logger");
 
 describe("createOauthPrivateApiClient", () => {
     it("should return private API client", () => {
