+++ test/unit/src/services/associationsService.getCompanyAssociations.test.ts
@@ -1,26 +1,29 @@
 import { Resource } from "@companieshouse/api-sdk-node";
-import { createOauthPrivateApiClient } from "../../../src/services/apiClientService";
-import { getUserAssociations } from "../../../src/services/associationsService";
-import { AssociationList, AssociationStatus } from "private-api-sdk-node/dist/services/associations/types";
+import { createOauthPrivateApiClient } from "../../../../src/services/apiClientService";
+import { getCompanyAssociations } from "../../../../src/services/associationsService";
+import { AssociationList } from "private-api-sdk-node/dist/services/associations/types";
 import { StatusCodes } from "http-status-codes";
-import { userAssociations } from "../../mocks/associations.mock";
+import { userAssociations } from "../../../mocks/associations.mock";
 import { Request } from "express";
 import { HttpError } from "http-errors";
-jest.mock("../../../src/services/apiClientService");
+
+jest.mock("../../../../src/services/apiClientService");
+jest.mock("../../../../src/lib/Logger");
 
 const mockCreateOauthPrivateApiClient = createOauthPrivateApiClient as jest.Mock;
-const mockSearchAssociations = jest.fn();
+const mockGetCompanyAssociations = jest.fn();
 
 mockCreateOauthPrivateApiClient.mockReturnValue({
     associationsService: {
-        searchAssociations: mockSearchAssociations
+        getCompanyAssociations: mockGetCompanyAssociations
     }
 });
 
 const reqest = {} as Request;
+const companyNumber = "NI038379";
 
 describe("associationsService", () => {
-    describe("getUserAssociations", () => {
+    describe("getCompanyAssociations", () => {
 
         beforeEach(() => {
             jest.clearAllMocks();
@@ -28,42 +31,41 @@ describe("associationsService", () => {
 
         it("should return associations", async () => {
             // Given
-            const status: AssociationStatus[] = [AssociationStatus.CONFIRMED];
             const sdkResource: Resource<AssociationList> = {
                 httpStatusCode: StatusCodes.OK,
                 resource: userAssociations
             };
-            mockSearchAssociations.mockResolvedValueOnce(sdkResource);
+            mockGetCompanyAssociations.mockResolvedValueOnce(sdkResource);
             // When
-            const result = await getUserAssociations(reqest, status);
+            const result = await getCompanyAssociations(reqest, companyNumber);
             // Then
             expect(result).toEqual(sdkResource.resource);
         });
 
         it("should return an error if no response returned from SDK", async () => {
-            mockSearchAssociations.mockResolvedValueOnce(undefined);
+            mockGetCompanyAssociations.mockResolvedValueOnce(undefined);
 
-            await expect(getUserAssociations(reqest, []))
+            await expect(getCompanyAssociations(reqest, companyNumber))
                 .rejects.toBe(undefined);
         });
 
         it("should throw an error if status code other than 200", async () => {
             const HTTP_STATUS_CODE = StatusCodes.BAD_REQUEST;
-            mockSearchAssociations.mockResolvedValueOnce({
+            mockGetCompanyAssociations.mockResolvedValueOnce({
                 httpStatusCode: HTTP_STATUS_CODE
             } as Resource<AssociationList>);
 
-            await expect(getUserAssociations(reqest, []))
+            await expect(getCompanyAssociations(reqest, companyNumber))
                 .rejects.toThrow(HttpError);
         });
 
         it("Should throw an error if no response resource returned from SDK", async () => {
             const HTTP_STATUS_CODE = StatusCodes.OK;
-            mockSearchAssociations.mockResolvedValueOnce({
+            mockGetCompanyAssociations.mockResolvedValueOnce({
                 httpStatusCode: HTTP_STATUS_CODE
             } as Resource<AssociationList>);
 
-            await expect(getUserAssociations(reqest, []))
+            await expect(getCompanyAssociations(reqest, companyNumber))
                 .rejects.toEqual({ httpStatusCode: StatusCodes.OK });
         });
     });
