+++ test/unit/src/services/associationsService.getUserAssociations.test.ts
@@ -1,28 +1,28 @@
 import { Resource } from "@companieshouse/api-sdk-node";
-import { createOauthPrivateApiClient } from "../../../src/services/apiClientService";
-import { getCompanyAssociations } from "../../../src/services/associationsService";
-import { AssociationList } from "private-api-sdk-node/dist/services/associations/types";
+import { createOauthPrivateApiClient } from "../../../../src/services/apiClientService";
+import { getUserAssociations } from "../../../../src/services/associationsService";
+import { AssociationList, AssociationStatus } from "private-api-sdk-node/dist/services/associations/types";
 import { StatusCodes } from "http-status-codes";
-import { userAssociations } from "../../mocks/associations.mock";
+import { userAssociations } from "../../../mocks/associations.mock";
 import { Request } from "express";
 import { HttpError } from "http-errors";
 
-jest.mock("../../../src/services/apiClientService");
+jest.mock("../../../../src/services/apiClientService");
+jest.mock("../../../../src/lib/Logger");
 
 const mockCreateOauthPrivateApiClient = createOauthPrivateApiClient as jest.Mock;
-const mockGetCompanyAssociations = jest.fn();
+const mockSearchAssociations = jest.fn();
 
 mockCreateOauthPrivateApiClient.mockReturnValue({
     associationsService: {
-        getCompanyAssociations: mockGetCompanyAssociations
+        searchAssociations: mockSearchAssociations
     }
 });
 
 const reqest = {} as Request;
-const companyNumber = "NI038379";
 
 describe("associationsService", () => {
-    describe("getCompanyAssociations", () => {
+    describe("getUserAssociations", () => {
 
         beforeEach(() => {
             jest.clearAllMocks();
@@ -30,41 +30,42 @@ describe("associationsService", () => {
 
         it("should return associations", async () => {
             // Given
+            const status: AssociationStatus[] = [AssociationStatus.CONFIRMED];
             const sdkResource: Resource<AssociationList> = {
                 httpStatusCode: StatusCodes.OK,
                 resource: userAssociations
             };
-            mockGetCompanyAssociations.mockResolvedValueOnce(sdkResource);
+            mockSearchAssociations.mockResolvedValueOnce(sdkResource);
             // When
-            const result = await getCompanyAssociations(reqest, companyNumber);
+            const result = await getUserAssociations(reqest, status);
             // Then
             expect(result).toEqual(sdkResource.resource);
         });
 
         it("should return an error if no response returned from SDK", async () => {
-            mockGetCompanyAssociations.mockResolvedValueOnce(undefined);
+            mockSearchAssociations.mockResolvedValueOnce(undefined);
 
-            await expect(getCompanyAssociations(reqest, companyNumber))
+            await expect(getUserAssociations(reqest, []))
                 .rejects.toBe(undefined);
         });
 
         it("should throw an error if status code other than 200", async () => {
             const HTTP_STATUS_CODE = StatusCodes.BAD_REQUEST;
-            mockGetCompanyAssociations.mockResolvedValueOnce({
+            mockSearchAssociations.mockResolvedValueOnce({
                 httpStatusCode: HTTP_STATUS_CODE
             } as Resource<AssociationList>);
 
-            await expect(getCompanyAssociations(reqest, companyNumber))
+            await expect(getUserAssociations(reqest, []))
                 .rejects.toThrow(HttpError);
         });
 
         it("Should throw an error if no response resource returned from SDK", async () => {
             const HTTP_STATUS_CODE = StatusCodes.OK;
-            mockGetCompanyAssociations.mockResolvedValueOnce({
+            mockSearchAssociations.mockResolvedValueOnce({
                 httpStatusCode: HTTP_STATUS_CODE
             } as Resource<AssociationList>);
 
-            await expect(getCompanyAssociations(reqest, companyNumber))
+            await expect(getUserAssociations(reqest, []))
                 .rejects.toEqual({ httpStatusCode: StatusCodes.OK });
         });
     });
