+++ test/unit/src/services/associationsService.postInvitation.test.ts
@@ -1,12 +1,13 @@
 import { Resource } from "@companieshouse/api-sdk-node";
-import { createOauthPrivateApiClient } from "../../../src/services/apiClientService";
-import { postInvitation } from "../../../src/services/associationsService";
+import { createOauthPrivateApiClient } from "../../../../src/services/apiClientService";
+import { postInvitation } from "../../../../src/services/associationsService";
 import { Errors, NewAssociationResponse } from "private-api-sdk-node/dist/services/associations/types";
 import { StatusCodes } from "http-status-codes";
-import { mockRequest } from "../../mocks/request.mock";
+import { mockRequest } from "../../../mocks/request.mock";
 import { BadRequest } from "http-errors";
 
-jest.mock("../../../src/services/apiClientService");
+jest.mock("../../../../src/services/apiClientService");
+jest.mock("../../../../src/lib/Logger");
 
 const mockCreateOauthPrivateApiClient = createOauthPrivateApiClient as jest.Mock;
 const mockPostInvitation = jest.fn();
