+++ test/unit/src/services/associationsService.removeUserFromCompanyAssociations.test.ts
@@ -1,12 +1,14 @@
 import { Request } from "express";
-import * as constants from "../../../src/constants";
-import * as associationsService from "../../../src/services/associationsService";
+import * as constants from "../../../../src/constants";
+import * as associationsService from "../../../../src/services/associationsService";
 import createError from "http-errors";
 import { StatusCodes } from "http-status-codes";
 
 const reqest = {} as Request;
 const updateAssociationStatusSpy = jest.spyOn(associationsService, "updateAssociationStatus");
 updateAssociationStatusSpy.mockImplementation(jest.fn());
+
+jest.mock("../../../../src/lib/Logger");
 
 describe("associationsService", () => {
     describe("isOrWasCompanyAssociatedWithUser", () => {
@@ -30,6 +32,7 @@ describe("associationsService", () => {
             const associationId = undefined;
             const expectedError = createError(StatusCodes.BAD_REQUEST, "Error on removal/cancellation: associtionId not provided");
             // When / Then
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
             await expect(associationsService.removeUserFromCompanyAssociations(reqest, associationId!))
                 .rejects.toEqual(expectedError);
         });
